requirements.txt
Django==4.1.3
django-crispy-forms==1.14.0
django-environ==0.9.0
djangorestframework==3.14.0
Pillow==9.4.0
